Add `UnversionedCaskChecker` helper class.
