+++ Library/Homebrew/cask/cask.rbi
@@ -2,6 +2,8 @@
 
 module Cask
   class Cask
+    def artifacts; end
+
     def homepage; end
   end
 end
