+++ manpages/brew.1
@@ -1250,7 +1250,7 @@ Check all casks with unversioned URLs in a given \fItap\fR for updates\.
 .
 .TP
 \fB\-n\fR, \fB\-\-dry\-run\fR
-List what would be done, but do not actually do anything\.
+Do everything except caching state and opening pull requests\.
 .
 .TP
 \fB\-\-limit\fR
@@ -1258,7 +1258,7 @@ Maximum runtime in minutes\.
 .
 .TP
 \fB\-\-state\-file\fR
-File for keeping track of state\.
+File for caching state\.
 .
 .SS "\fBcat\fR \fIformula\fR|\fIcask\fR"
 Display the source of a \fIformula\fR or \fIcask\fR\.
